requirements.txt
# Windows-friendly pinned deps for invoice classification (Python 3.13)
numpy==2.1.3
pypdf==4.3.1
pillow==10.4.0
imagehash==4.3.1
pdf2image==1.17.0
pytesseract==0.3.13
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0
PyYAML==6.0.2
tqdm==4.66.4
